Added Makefile
By default the helloworld.py will be installed in:
/usr/local/bin
with execution permissions.

If PREFIX is set it will be using the environment PREFIX value.
Thanks to the DESTDIR variable it is already prepared to be used by Debian.
If it was not it would have to be modified as described in:

https://www.debian.org/doc/manuals/maint-guide/modify.en.html#destdir